refactor: Refactor record panel and notice handling
